requirements.txt
celery==5.2.7
redis==4.3.3
click==8.1.3
Pillow==9.1.1
opencv-python==4.5.5.64
imutils==0.5.4
python-dotenv==0.20.0
psycopg2-binary
boto3==1.28.68